+++ apps/assets/src/components/asset/table/components/SubRowContent.tsx
@@ -238,8 +238,10 @@ export const SubRowContent = ({
             createConvertButton()}
           {/* to withdraw axelar, the users need to convert
           the erc20 to ibc-balance */}
-          {item.symbol !== EVMOS_SYMBOL &&
-            item.handledByExternalUI !== null &&
+          {((item.symbol !== EVMOS_SYMBOL &&
+            item.handledByExternalUI !== null) ||
+            (item.symbol !== EVMOS_SYMBOL &&
+              item.chainIdentifier?.toLowerCase() === "gravity")) &&
             createConvertButton()}
         </div>
       </div>